+++ addOns/network/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Use always a plain connection to the outgoing HTTP proxy (Issue 7594).
 - Do not change the case of the Content-Length header.
+- Use the available response content when the Content-Length is more than what is available.
 
 ## [0.5.0] - 2022-11-09
 ### Fixed